Add authcheck endpoint
used with nginx auth_request, which is now serving
all the files directly, instead of the download() endpoint
in BaseController. This trades efficiency of serving (nginx
is a really good file server) with the little download counter
in mysql thingy that download() does. I can just scrape the
logs for that though.